Parcial 2/requirements.txt
mysql-connector-python



